feat: Add Manager class for running tests
The Manager class is responsible for running tests in the game. This commit adds a new file, Manager.js, which contains the implementation of the Manager class. The class has a constructor that takes a game object as a parameter and a run() method for running the tests. The run() method logs a message to the console and returns true.